# Fix parallel formatting and progress flicker in clickhouse-client