--- a/src/Client/ClientBase.h
+++ b/src/Client/ClientBase.h
@@ -155,6 +155,7 @@ protected:
     ConnectionParameters connection_parameters;
 
     String format; /// Query results output format.
+    bool select_into_file = false; /// If writing result INTO OUTFILE. It affects progress rendering.
     bool is_default_format = true; /// false, if format is set in the config or command line.
     size_t format_max_block_size = 0; /// Max block size for console output.
     String insert_format; /// Format of INSERT data that is read from stdin in batch mode.